# docs/swagger.yaml
basePath: /v1
definitions:
  handler.categoryResponse:
    properties:
      id:
        example: ed6ac028-8016-4cbd-aeee-c3a155cdb2a4
        type: string
      name:
        example: Lanche
        type: string
    type: object
  handler.clientResponse:
    properties:
      email:
        example: john-doe@email.com
        type: string
      id:
        example: ed6ac028-8016-4cbd-aeee-c3a155cdb2a4
        type: string
      name:
        example: John Doe
        type: string
    type: object
  handler.createClientRequest:
    properties:
      email:
        example: john-doe@email.com
        type: string
      name:
        example: John Doe
        type: string
    required:
    - email
    - name
    type: object
  handler.createProductRequest:
    properties:
      category_id:
        example: ed6ac028-8016-4cbd-aeee-c3a155cdb2a4
        minLength: 1
        type: string
      description:
        example: Lanche com batata
        type: string
      image:
        example: https://
        type: string
      name:
        example: Lanche
        type: string
      value:
        example: 10.9
        type: number
    required:
    - name
    - value
    type: object
  handler.errorResponse:
    properties:
      messages:
        example:
        - Error message 1
        - ' Error message 2'
        items:
          type: string
        type: array
      success:
        example: false
        type: boolean
    type: object
  handler.productResponse:
    properties:
      category:
        $ref: '#/definitions/handler.categoryResponse'
      created_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      description:
        example: Lanche com bacon
        type: string
      id:
        example: ed6ac028-8016-4cbd-aeee-c3a155cdb2a4
        type: string
      image:
        example: https://
        type: string
      name:
        example: Lanche 1
        type: string
      updated_at:
        example: "1970-01-01T00:00:00Z"
        type: string
      value:
        example: 10.9
        type: number
    type: object
  handler.updateProductRequest:
    properties:
      category_id:
        example: ed6ac028-8016-4cbd-aeee-c3a155cdb2a4
        minLength: 1
        type: string
      description:
        example: Lanche com batata
        type: string
      image:
        example: https://
        type: string
      name:
        example: Lanche
        type: string
      value:
        example: 10.9
        type: number
    required:
    - name
    - value
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API em Go para o desafio na pos-tech fiap de Software Architecture.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: POS-Tech API
  version: "1.0"
paths:
  /clients:
    post:
      consumes:
      - application/json
      description: Registra um novo cliente com nome e e-mail
      parameters:
      - description: Registrar novo cliente request
        in: body
        name: createClientRequest
        required: true
        schema:
          $ref: '#/definitions/handler.createClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Cliente registrado
          schema:
            $ref: '#/definitions/handler.clientResponse'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Registra um novo cliente
      tags:
      - Clients
  /clients/{cpf}:
    get:
      consumes:
      - application/json
      description: buscar um cliente pelo Cpf
      parameters:
      - description: CPF
        in: path
        name: cpf
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Cliente
          schema:
            $ref: '#/definitions/handler.clientResponse'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Cliente nao encontrado
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Busca um cliente
      tags:
      - Clients
  /products:
    get:
      consumes:
      - application/json
      description: Lista os produtos podendo buscar por categoria
      parameters:
      - description: Id da categoria
        in: query
        name: category_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Produtos listados
          schema:
            items:
              $ref: '#/definitions/handler.productResponse'
            type: array
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "500":
          description: Erro interno
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Lista os produtos
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: registra um novo produto
      parameters:
      - description: Registrar novo produto body
        in: body
        name: createProductRequest
        required: true
        schema:
          $ref: '#/definitions/handler.createProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Produto registrado
          schema:
            $ref: '#/definitions/handler.productResponse'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Registra um novo produto
      tags:
      - Products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Remove um produto por meio de seu identificador
      parameters:
      - description: Id do produto
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Produto removido
          schema:
            $ref: '#/definitions/handler.productResponse'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Produto nao encontrado
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Remove um produto
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Atualiza um produto
      parameters:
      - description: Id do produto
        in: path
        name: id
        required: true
        type: string
      - description: Atualizar produto body
        in: body
        name: updateProductRequest
        required: true
        schema:
          $ref: '#/definitions/handler.updateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Produto atualizado
          schema:
            $ref: '#/definitions/handler.productResponse'
        "400":
          description: Erro de validação
          schema:
            $ref: '#/definitions/handler.errorResponse'
        "404":
          description: Produto nao encontrado
          schema:
            $ref: '#/definitions/handler.errorResponse'
      summary: Atualiza um produto
      tags:
      - Products
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"